@@ -1,10 +1,10 @@
-import AgoraRTC from "agora-rtc-sdk-ng";
-import { useEffect, useRef, useState } from "react";
-import { useParams } from "react-router";
-import generateToken from "../createToken";
-import { generateUid } from "../uidGen";
-import Auth from "./Auth";
-import LiveChat from "./LiveChat";
+import AgoraRTC from 'agora-rtc-sdk-ng';
+import { useEffect, useRef, useState } from 'react';
+import { useParams } from 'react-router';
+import generateToken from '../createToken';
+import { generateUid } from '../uidGen';
+import Auth from './Auth';
+import LiveChat from './LiveChat';
 
 const rtc = {
 	// For the local audio and video tracks.
@@ -14,33 +14,28 @@ const rtc = {
 };
 
 const options = {
-	appId: "7ac8167595aa47aeb4ddf6b34353ec38",
+	appId: '7ac8167595aa47aeb4ddf6b34353ec38',
 	uid: generateUid(),
 };
 
-rtc.client = AgoraRTC.createClient({ mode: "live", codec: "vp8" });
+rtc.client = AgoraRTC.createClient({ mode: 'live', codec: 'vp8' });
 
-const WatchSream = ({ user }) => {
+const WatchStream = ({ user }) => {
 	const { streamId } = useParams();
 	const [showStream, setShowStream] = useState(false);
 	const videoStream = useRef();
 
 	useEffect(() => {
 		const getStream = async () => {
-			rtc.client.setClientRole("audience");
-			await rtc.client.join(
-				options.appId,
-				"test",
-				generateToken(options.uid, "test"),
-				options.uid
-			);
+			rtc.client.setClientRole('audience');
+			await rtc.client.join(options.appId, 'test', generateToken(options.uid, 'test'), options.uid);
 
-			rtc.client.on("user-published", async (user, mediaType) => {
+			rtc.client.on('user-published', async (user, mediaType) => {
 				// Subscribe to a remote user.
 				await rtc.client.subscribe(user, mediaType);
-				console.log("subscribe success");
+				console.log('subscribe success');
 
-				if (mediaType === "video") {
+				if (mediaType === 'video') {
 					// show video stream
 					setShowStream(true);
 
@@ -52,7 +47,7 @@ const WatchSream = ({ user }) => {
 				}
 
 				// If the subscribed track is audio.
-				if (mediaType === "audio") {
+				if (mediaType === 'audio') {
 					// Get `RemoteAudioTrack` in the `user` object.
 					const remoteAudioTrack = user.audioTrack;
 					// Play the audio track. No need to pass any DOM element.
@@ -71,15 +66,15 @@ const WatchSream = ({ user }) => {
 
 	return (
 		<>
-			<header className="streaming-pages-header">
+			<header className='streaming-pages-header'>
 				<h1>Watch stream</h1>
 				<Auth user={user} />
 			</header>
-			<div className="streamContainer">
+			<div className='streamContainer'>
 				{!showStream ? (
-					<div className="no-video">No Video Stream</div>
+					<div className='no-video'>No Video Stream</div>
 				) : (
-					<div className="video-container" ref={videoStream}></div>
+					<div className='video-container' ref={videoStream}></div>
 				)}
 				<LiveChat streamId={streamId} user={user} />
 			</div>
@@ -87,4 +82,4 @@ const WatchSream = ({ user }) => {
 	);
 };
 
-export default WatchSream;
+export default WatchStream;
